Explicitly cast FIONBIO constant to int to suppress compiler warning on some platforms.

+++ asio/include/asio/detail/io_control.hpp
@@ -45,7 +45,7 @@ public:
   // Get the name of the IO control command.
   int name() const
   {
-    return FIONBIO;
+    return static_cast<int>(FIONBIO);
   }
 
   // Set the value of the I/O control command.
@@ -95,7 +95,7 @@ public:
   // Get the name of the IO control command.
   int name() const
   {
-    return FIONREAD;
+    return static_cast<int>(FIONREAD);
   }
 
   // Set the value of the I/O control command.
